Added lots of new stuff
Added an Activity which checks if the user is already registered or not.
If he isn't or his password doesn't work, then he will be automaticly
redirected to the RegisterActivity. If everything goes well, then he
will be redirected to the Conversations activity.

Note that this hasn't be tested, so it may not work properly.