Very important security fix for verification
There was a problem when registering your device for WhatsApi. All
devices had the same id, that means, if I verified my number on
WhatsApi, somebody else could register my number too on his phone
without requesting a sms.

I fixed it by creating a random 20 bytes array, which means that the
risk of the same id is much smaller, compared to before.